Merge the Q&A sessions into one for Day#2
The Q&A session require a special type of meeting with
the ability of the audience to participate and engage more
actively.

Having a single Q&A session at the end allows simplifying the
overall logistics.

Also extend Tony's talk about AWS-Gerrit and X-Ray to 30 mins
so that he can have more time to explain what it is and how
we also use it for E2E tests for Gerrit releases.

Change-Id: Ia38311da42a3b5a199d74d3667d256fff80571da